refactor: rename "config" file to "virtual" (#10638)
* refactor: rename "config" file to "virtual"

* chore: changeset

* fix(test): config import
