+++ .changeset/plenty-owls-wait.md
@@ -1,0 +1,5 @@
+---
+"@astrojs/db": minor
+---
+
+Rename internal import from `runtime/config` to `runtime/virtual`
