+++ packages/db/test/unit/reset-queries.test.js
@@ -3,7 +3,7 @@ import { describe, it } from 'mocha';
 import { getMigrationQueries } from '../../dist/core/cli/migration-queries.js';
 import { MIGRATION_VERSION } from '../../dist/core/consts.js';
 import { tableSchema } from '../../dist/core/schemas.js';
-import { column, defineTable } from '../../dist/runtime/config.js';
+import { column, defineTable } from '../../dist/runtime/virtual.js';
 
 const TABLE_NAME = 'Users';
