+++ packages/db/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig.base.json",
   "include": ["src"],
-  "exclude": ["src/runtime/config.ts"],
+  "exclude": ["src/runtime/virtual.ts"],
   "compilerOptions": {
     "outDir": "./dist"
   }
